requirements.txt
python-dotenv
openai
requests
pillow
